+++ social_core/backends/shopify.py
@@ -22,7 +22,7 @@ class ShopifyOAuth2(BaseOAuth2):
     REDIRECT_STATE = False
 
     @property
-    def shopifyAPI(self):
+    def shopify_api(self):
         if not hasattr(self, '_shopify_api'):
             fp, pathname, description = imp.find_module('shopify')
             self._shopify_api = imp.load_module('shopify', fp, pathname,
@@ -42,7 +42,8 @@ class ShopifyOAuth2(BaseOAuth2):
         extra_data field"""
         data = super(ShopifyOAuth2, self).extra_data(user, uid, response,
                                                      details, *args, **kwargs)
-        session = self.shopifyAPI.Session(self.data.get('shop').strip())
+        session = self.shopify_api.Session(self.data.get('shop').strip(),
+                                           version='2019-04')
         # Get, and store the permanent token
         token = session.request_token(data['access_token'])
         data['access_token'] = token
@@ -50,12 +51,13 @@ class ShopifyOAuth2(BaseOAuth2):
 
     def auth_url(self):
         key, secret = self.get_key_and_secret()
-        self.shopifyAPI.Session.setup(api_key=key, secret=secret)
+        self.shopify_api.Session.setup(api_key=key, secret=secret)
         scope = self.get_scope()
         state = self.state_token()
         self.strategy.session_set(self.name + '_state', state)
         redirect_uri = self.get_redirect_uri(state)
-        session = self.shopifyAPI.Session(self.data.get('shop').strip())
+        session = self.shopify_api.Session(self.data.get('shop').strip(),
+                                           version='2019-04')
         return session.create_permission_url(
             scope=scope,
             redirect_uri=redirect_uri
@@ -69,10 +71,12 @@ class ShopifyOAuth2(BaseOAuth2):
         key, secret = self.get_key_and_secret()
         try:
             shop_url = self.data.get('shop')
-            self.shopifyAPI.Session.setup(api_key=key, secret=secret)
-            shopify_session = self.shopifyAPI.Session(shop_url, self.data)
+            self.shopify_api.Session.setup(api_key=key, secret=secret)
+            shopify_session = self.shopify_api.Session(shop_url,
+                                                       version='2019-04',
+                                                       token=self.data)
             access_token = shopify_session.token
-        except self.shopifyAPI.ValidationException:
+        except self.shopify_api.ValidationException:
             raise AuthCanceled(self)
         else:
             if not access_token: